+++ src/childBridgeService.ts
@@ -627,6 +627,7 @@ export class ChildBridgeService {
     const bridgeOptions: BridgeOptions = {
       cachedAccessoriesDir: User.cachedAccessoryPath(),
       cachedAccessoriesItemName: `cachedAccessories.${this.bridgeConfig.username.replace(COLON_RE, '').toUpperCase()}`,
+      externalAccessoriesItemName: `externalAccessories.${this.bridgeConfig.username.replace(COLON_RE, '').toUpperCase()}`,
     }
 
     // shallow copy the homebridge options to the bridge options object